requirements.txt
streamlit==1.24.1

sentence-transformers==2.2.2
torch==2.0.1
faiss-cpu==1.7.4
PyPDF2==3.0.1

requests~=2.32.3
python-dotenv~=1.1.0
langchain~=0.3.25
feedparser~=6.0.11
langchain-community~=0.3.23